refactor(approval-gate): extract register.rs from lib.rs
Final step of breaking lib.rs into focused modules. This commit
lifts the iii function/trigger wiring — the heart of the worker
startup path — into its own module.

Moved (lib.rs -> register.rs):
- register() (~420 lines: the subscriber closure + 8 function
  registrations + trigger registration + sweeper spawn)
- Refs struct (the handle bag returned to the binary)
- FN_RESOLVE / FN_LIST_PENDING / FN_LIST_UNDELIVERED /
  FN_CONSUME_UNDELIVERED / FN_ACK_DELIVERED / FN_FLUSH_DELIVERED /
  FN_SWEEP_SESSION / FN_LOOKUP_RECORD constants
- STATE_SCOPE constant

lib.rs is now ~50 lines of module declarations + re-exports +
~2580 lines of inline tests. The test mod uses super::* plus
local serde_json and std::sync imports.

All public items stay re-exported from the crate root so the
binary's `use approval_gate::register;` keeps working unchanged.
Pure move, no behavior change.

141 approval-gate tests pass; turn-orchestrator + harness + shell
verified green (turn-orchestrator's pre-existing dual_write flake
notwithstanding).
